move models/clients.py to services/aws_clients.py for clarity
