+++ api/main.py
@@ -5,7 +5,7 @@ load_dotenv()
 
 from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from models.clients import AWSClients
+from services.aws_clients import AWSClients
 from routes import router
 
 class App(FastAPI):
